Moved C functions to separate folder.
