+++ API/native_libs/src/LifeTimeManger.cpp
@@ -9,8 +9,3 @@ LifetimeManager::LifetimeManager()
 LifetimeManager::~LifetimeManager()
 {
 }
-
-void release(void *handle) noexcept
-{
-    LifetimeManager::instance().releaseOwnership(handle);
-}
