+++ API/native_libs/src/c_wrappers/memory.cpp
@@ -1,0 +1,7 @@
+#include <API/native_libs/src/LifeTimeManager.h>
+#include "memory.h"
+
+void release(void *handle) noexcept
+{
+    LifetimeManager::instance().releaseOwnership(handle);
+}
