+++ API/native_libs/src/c_wrappers/memory.h
@@ -1,0 +1,6 @@
+#ifndef FFITESTHELPER_MEMORY_H
+#define FFITESTHELPER_MEMORY_H
+
+extern "C" void release(void *handle) noexcept;
+
+#endif //FFITESTHELPER_MEMORY_H
