+++ API/native_libs/src/ops/BinaryOperation.cpp
@@ -3,8 +3,3 @@
 //
 
 #include "BinaryOperation.h"
-
-Operation* make_op_binary(const char* name, Operation* a, Operation* b) {
-    std::string cpp_name(name);
-    return not_implemented<Operation*>();
-}
