+++ API/native_libs/src/ops/Const.cpp
@@ -3,7 +3,3 @@
 //
 
 #include "Const.h"
-
-Operation* make_op_const(TODOType* tensor) {
-    return not_implemented<Operation*>();
-}
